@@ -1,24 +1,27 @@
-var React = require('react');
-var ReactDom = require('react-dom');
-var ProgressBar = require('../src/main.js');
-var Circle = ProgressBar.Circle;
+import React, { Component } from 'react';
+import { render } from 'react-dom';
+import { Circle } from '../src/main.js';
 
-var App = React.createClass({
-    getInitialState: function() {
-        return {
+class App extends Component {
+    constructor(props) {
+        super(props);
+
+        this.state = {
             progress: 0.1,
             options: {}
         };
-    },
+    }
 
-    render: function() {
+    render() {
         return <Circle initialAnimate={this.state.initialAnimate} options={this.state.options} progress={this.state.progress} />;
-    },
+    }
 
-    componentDidMount: function() {
-        var self = this;
+    componentDidMount() {
+        const self = this;
+        
         setTimeout(function() {
             console.log('setstate')
+            
             self.setState({
                 progress: 1
             });
@@ -26,16 +29,16 @@ var App = React.createClass({
 
         setTimeout(function() {
             console.log('setstate')
+            
             self.setState({
                 initialAnimate: true,
                 progress: 1
             });
         }, 500);
-
     }
-});
+};
 
-ReactDom.render(
+render(
     <App />,
-    document.querySelector('body')
+    document.querySelector('#progress')
 );
