@@ -1,61 +1,62 @@
-var React = require('react');
-var ReactDom = require('react-dom');
-var isEqual = require('lodash.isequal');
-var ProgressBar = require('progressbar.js');
+import React, { Component } from 'react';
+import { findDOMNode } from 'react-dom';
+import isEqual from 'lodash/isEqual';
+import ProgressBar from 'progressbar.js';
 
-var Shape = React.createClass({
-    getDefaultProps: function getDefaultProps() {
-        return {
-            ShapeClass: null,
-            options: {},
-            progress: 0,
-            text: null,
-            initialAnimate: false,
-            containerStyle: {},
-            containerClassName: '.progressbar-container'
-        };
-    },
+class Shape extends Component {
+    defaultProps: {
+        ShapeClass: null,
+        options: {},
+        progress: 0,
+        text: null,
+        initialAnimate: false,
+        containerStyle: {},
+        containerClassName: '.progressbar-container'
+    }
 
-    getInitialState: function getInitialState() {
-        return {
+    constructor(props) {
+        super(props);
+
+        this.state = {
             shape: null
         };
-    },
+    }
 
-    render: function render() {
-        var style = this.props.containerStyle;
-        var className = this.props.containerClassName;
+    render() {
+        const {style, className} = this.props;
 
         return <div className={className} style={style} ref="progressBar"></div>;
-    },
+    }
 
-    componentWillReceiveProps: function componentWillReceiveProps(nextProps) {
+    componentWillReceiveProps(nextProps) {
         if (!isEqual(this.props.options, nextProps.options)) {
             this._destroy();
             this._create(nextProps, this.props);
+
             return;
         }
 
         this._animateProgress(nextProps.progress);
         this._setText(nextProps.text);
-    },
+    }
 
-    componentDidMount: function componentDidMount() {
+    componentDidMount() {
         this._create(this.props);
-    },
+    }
 
-    componentWillUnmount: function componentWillUnmount() {
-        this._destroy()
-    },
+    componentWillUnmount() {
+        this._destroy();
+    }
 
-    _create: function _create(props, oldProps) {
+    _create(props, oldProps) {
         if (this.state.shape !== null) {
             throw new Error('Progressbar is already created');
         }
 
         // setState function is not used to prevent a new render cycle
         // This handling happens outside of React component's lifecycle
-        var container = ReactDom.findDOMNode(this.refs.progressBar);
+        var container = findDOMNode(this.refs.progressBar);
+
         this.state.shape = new props.ShapeClass(
             container,
             props.options
@@ -72,47 +73,47 @@ var Shape = React.createClass({
         }
 
         this._setText(props.text);
-    },
+    }
 
-    _destroy: function _destroy() {
+    _destroy() {
         if (this.state.shape) {
             this.state.shape.destroy();
             this.state.shape = null;
         }
-    },
+    }
 
-    _animateProgress: function _animateProgress(progress) {
+    _animateProgress(progress) {
         this.state.shape.animate(progress);
-    },
+    }
 
-    _setProgress: function _setProgress(progress) {
+    _setProgress(progress) {
         this.state.shape.set(progress);
-    },
+    }
 
-    _setText: function _setText(text) {
+    _setText(text) {
         if (text) {
             this.state.shape.setText(text);
         }
     }
-});
+};
 
-var Line = React.createClass({
+class Line extends Component {
     render() {
         return <Shape {...this.props} ShapeClass={ProgressBar.Line} />;
     }
-});
+};
 
-var Circle = React.createClass({
+class Circle extends Component {
     render() {
         return <Shape {...this.props} ShapeClass={ProgressBar.Circle} />;
     }
-});
+};
 
-var SemiCircle = React.createClass({
+class SemiCircle extends Component {
     render() {
         return <Shape {...this.props} ShapeClass={ProgressBar.SemiCircle} />;
     }
-});
+};
 
 module.exports = {
     Line: Line,
